% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{My Example Dataset}
\format{
A data frame with details.
}
\source{
Artificially generated based on authors experience.
}
\usage{
data(example_data)
}
\description{
This is an example dataset included in the package.
}
\examples{
data(example_data)
}
\keyword{datasets}
