% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{generate_well_plate_template}
\alias{generate_well_plate_template}
\title{A function to generate plate layouts}
\usage{
generate_well_plate_template(plate_type, reserve_last_column = TRUE)
}
\arguments{
\item{plate_type}{The type of the plate to generate. Could be: "24w", "48w", "96w", "384w"}

\item{reserve_last_column}{Optional. Defaults to TRUE. Last column is reserved for control samples or something else}
}
\value{
Character matrix resembling specified plate
}
\description{
`generate_well_plate_template()` This function is used to generate commonly used well plate templates in 
life sciences. Supports "24 well", "48 well", "96 well", "384 well" layouts. Last columns of the plate
could be reserved
}
\details{
Generates a layout matrix based on the specified format and if the last column should be reserved (NA values).
}
\examples{

plate_matrix = generate_well_plate_template(plate_type = "48w")
print(plate_matrix)
}
