% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{\%!in\%}
\alias{\%!in\%}
\title{A custom infix operator performing "not in" action}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{A vector of objects.}

\item{y}{A target vector to compare to.}
}
\value{
Boolean vector
}
\description{
An operator that performs an inverse action of %in%
}
\details{
Returns TRUE if element is not within the target vector
}
\examples{

a = 1:10
b = 1:20
a \%!in\% b
}
