% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{plate_layout_randomizer}
\alias{plate_layout_randomizer}
\title{A function to generate random layout for a dataset with samples}
\usage{
plate_layout_randomizer(
  df_all_studies,
  column_study,
  column_sample_type,
  column_sample_ID,
  column_conf,
  plate_template
)
}
\arguments{
\item{df_all_studies}{A dataframe containing information on Sample ID, Study, Type, Confounder (should not be named as "Confounder")}

\item{column_study}{A number indicating a column with study of origin}

\item{column_sample_type}{A number indicating a column with sample types}

\item{column_sample_ID}{A number indicating a column with sample IDs}

\item{column_conf}{A number indicating a column with confounders}

\item{plate_template}{A matrix where each cell has a character name, and rows and columns are named. Cells with NA will be skipped}
}
\value{
A list with randomized samples
}
\description{
`plate_layout_randomizer()` This function is used to generate randomized layouts for specified plates
}
\details{
Randomly distributes samples on the specified plate format.Tries to allocate samples from same study and with the same confounder on one plate.
Distribution of samples may be sub-optimal. If sizes of studies are less than size of a plate and empty cells are not an issue, the
best practice is to use randomization per study and not mix samples from different studies.
}
\examples{

data("example_data")
dataset = example_data
head(dataset)
plate_template = generate_well_plate_template(plate_type = "96w", 
reserve_last_column = TRUE)
plate_template
test_randomization = plate_layout_randomizer(df_all_studies = dataset,  
column_study = 5, 
column_sample_type = 2, 
column_sample_ID = 1, 
column_conf = 7, 
plate_template = plate_template)
test_randomization
}
