% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{randomizer_main}
\alias{randomizer_main}
\title{A function to generate random layout for a dataset with samples and visualize variable across plates}
\usage{
randomizer_main(
  df_studies,
  col_num_study,
  col_num_sample_ID,
  col_num_sample_type,
  col_num_conf,
  column_numbers_vis,
  plate_template,
  randomiz_number = 10,
  main_dir = "randomiz_vis",
  dir_combined_img = "randomiz_vis_combined_img"
)
}
\arguments{
\item{df_studies}{A dataframe containing information on Sample ID, Study, Type, Confounder (should not be named as "Confounder")}

\item{col_num_study}{A number indicating a column with study of origin}

\item{col_num_sample_ID}{A number indicating a column with sample IDs}

\item{col_num_sample_type}{A number indicating a column with sample types}

\item{col_num_conf}{A number indicating a column with confounders}

\item{column_numbers_vis}{A numeric vector of integers indicating columns with variables for visualization}

\item{plate_template}{A matrix where each cell has a character name, and rows and columns are named. Cells with NA will be skipped}

\item{randomiz_number}{Number of randomizations to perform (default is 10)}

\item{main_dir}{The path to directory for combined images (default: "randomiz_vis_combined_img")}
}
\value{
Nothing. All outputs are saved in the specified directories
}
\description{
`randomizer_main()` This function is used to generate randomized layouts for specified plates specified number of times. Saves
outputs as excel. Saves visualizations of variables for each layout
}
\details{
Randomly distributes samples on the specified plate format.Tries to allocate samples from same study and with the same confounder on one plate.
Distribution of samples may be sub-optimal. If sizes of studies are less than size of a plate and empty cells are not an issue, the
best practice is to use randomization per study and not mix samples from different studies. 

Generates visualization of layouts by specified columns in column_numbers_vis
}
\examples{

data("example_data")
dataset = example_data
head(dataset)
plate_template = generate_well_plate_template(plate_type = "96w", 
reserve_last_column = TRUE)
plate_template
randomizer_main(df_studies = dataset,  
col_num_study = 5, 
col_num_sample_type = 2, 
col_num_sample_ID = 1, 
col_num_conf = 7, 
column_numbers_vis = c(2,7,3,4,5),
plate_template = plate_template)
}
