% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{visualize_table_layout}
\alias{visualize_table_layout}
\title{A function to download trials for a set of diseases and drugs}
\usage{
visualize_table_layout(
  table,
  condtion = NULL,
  levels_var = NULL,
  file_name_full,
  table_label,
  gwidth = 100,
  gheight = 35
)
}
\arguments{
\item{table}{A table matrix to visualize}

\item{condtion}{Optional. Character expression for a threshold value in a form ">5" or "==3"}

\item{levels_var}{Optional. Character vector of levels for a categorical variable}

\item{file_name_full}{A full path to store figure}

\item{table_label}{A character name for the table}

\item{gwidth}{Width of a column. Default is 100}

\item{gheight}{Height of a column. Default is 35}
}
\value{
Nothing is returned. The function is expected to create an image in the specified path
}
\description{
`visualize_table_layout()` This function is used to visualize table layout based on the specified variable
}
\details{
Generated PNG image showing distribution of samples on the plate based on the specified variable and its 
threshold if needed. Expected to be used internally. DOES NOT WORK IF INPUT MATRIX HAS NO NAMES FOR ROWS
AND COLUMNS
}
\examples{

matrix_sample = 1:16
matrix_sample = matrix(matrix_sample, ncol=4, nrow=4)
rownames(matrix_sample) = LETTERS[1:4]
colnames(matrix_sample) = as.character(1:4)
# All values above 5 will be in a different color
visualize_table_layout(table = matrix_sample, 
                       condtion = ">5",
                       file_name_full = "test.png", 
                       table_label = "Test table")
}
